refactor(__init__.py): simplified funtion path
