+++ src/orquestra/vqa/algorithms/__init__.py
@@ -1,0 +1,3 @@
+from .qaoa import QAOA
+from .qcbm import QCBM
+from .vqe import VQE
