+++ src/orquestra/vqa/algorithms/qaoa.py
@@ -14,9 +14,9 @@ from orquestra.quantum.estimation import (
 from orquestra.quantum.operators import PauliRepresentation
 from scipy.optimize import OptimizeResult
 
-from orquestra.vqa.ansatz.qaoa_farhi import QAOAFarhiAnsatz
+from orquestra.vqa.ansatz import QAOAFarhiAnsatz
 from orquestra.vqa.api.ansatz import Ansatz
-from orquestra.vqa.cost_function.cost_function import (
+from orquestra.vqa.cost_function import (
     create_cost_function,
     substitution_based_estimation_tasks_factory,
 )
