+++ src/orquestra/vqa/ansatz/qcbm/__init__.py
@@ -1,4 +1,4 @@
 ################################################################################
 # © Copyright 2022 Zapata Computing Inc.
 ################################################################################
-from ._qcbm import QCBMAnsatz
+# from ._qcbm import QCBMAnsatz
