+++ src/orquestra/vqa/cost_function/__init__.py
@@ -1,0 +1,10 @@
+from .cost_function import (
+    add_normal_noise,
+    create_cost_function,
+    dynamic_circuit_estimation_tasks_factory,
+    expectation_value_estimation_tasks_factory,
+    fix_parameters,
+    substitution_based_estimation_tasks_factory,
+    sum_expectation_values,
+)
+from .qcbm_cost_function import create_QCBM_cost_function
