+++ tests/orquestra/vqa/opt/fourier_test.py
@@ -8,13 +8,13 @@ from orquestra.opt.api.functions import CallableWithGradient, function_with_grad
 from orquestra.opt.api.optimizer import optimization_result
 from orquestra.opt.api.optimizer_test import NESTED_OPTIMIZER_CONTRACTS
 from orquestra.opt.mock_objects import MockOptimizer, mock_cost_function
-from orquestra.quantum.estimation._estimation import calculate_exact_expectation_values
+from orquestra.quantum.estimation import calculate_exact_expectation_values
 from orquestra.quantum.operators import PauliTerm
 from orquestra.quantum.runners import SymbolicSimulator
 
-from orquestra.vqa.ansatz.qaoa_farhi import QAOAFarhiAnsatz
+from orquestra.vqa.ansatz import QAOAFarhiAnsatz
 from orquestra.vqa.api.ansatz import Ansatz
-from orquestra.vqa.cost_function.cost_function import (
+from orquestra.vqa.cost_function import (
     create_cost_function,
     substitution_based_estimation_tasks_factory,
 )
